+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Referendums list`, `External list`, `Proposals list` on `Democracy page`
 - Progress info block
 - `Proposals list` on `Treasury overview tab`
+- Treasury summary
 
 #### Fixed
 - Collected winning data on auction tab